[COMMON] soc: samsung: cpif: fix prevent issue
CID 277217, 276710, 310329, 277446

+++ drivers/soc/samsung/cpif/bootdump_io_device.c
@@ -334,7 +334,7 @@ static long bootdump_ioctl(struct file *filp, unsigned int cmd, unsigned long ar
 		switch (ld->protocol) {
 		case PROTOCOL_SIPC:
 			if (arg)
-				ld->crash_reason.type = arg;
+				ld->crash_reason.type = (u32)arg;
 			mif_err("%s: IOCTL_TRIGGER_CP_CRASH (%lu)\n",
 					iod->name, arg);
 			break;
